+++ src/info.py
@@ -1,4 +1,4 @@
 class Info:
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.10b9" # see https://peps.python.org/pep-0440/
-    AppDateStr = "2023-03-06"
+    AppVersionStr = "0.10b10" # see https://peps.python.org/pep-0440/
+    AppDateStr = "2023-03-15"